# api/src/app/api/collections/[id]/items/route.ts
import { NextRequest, NextResponse } from "next/server";
import { CORS_HEADERS, corsPreflightResponse } from "@/lib/cors";

export const runtime = "edge";

/**
 * OGC API - Features items endpoint.
 *
 * Returns GeoJSON features for a given collection with support for:
 * - bbox filtering
 * - pagination (limit/offset)
 * - CQL2-like filtering (simple property filters)
 *
 * Follows OGC API - Features 1.0 specification.
 */

// Data source mapping: collection ID -> upstream GeoJSON URL
// External URLs are absolute; internal paths are resolved relative to the request origin
// Partial keeps the lookup result possibly undefined: collection ids come straight
// from the URL, so a miss is a normal runtime case that must stay reachable.
const COLLECTION_SOURCES: Partial<Record<string, { url: string; cacheTtl: number }>> = {
  earthquakes: {
    url: "https://earthquake.usgs.gov/earthquakes/feed/v1.0/summary/all_day.geojson",
    cacheTtl: 60,
  },
  natural_events: {
    url: "https://eonet.gsfc.nasa.gov/api/v3/events/geojson?status=open&limit=200",
    cacheTtl: 300,
  },
  wildfires: {
    url: "/api/wildfires",
    cacheTtl: 3600,
  },
  nlnog_nodes: {
    url: "/api/nlnog",
    cacheTtl: 3600,
  },
  warnings: {
    url: "/api/weather/warnings",
    cacheTtl: 120,
  },
  waterways: {
    // bbox form: /api/waterways only accepts bbox (lat/lon/radius 400s) —
    // ~50km around New York, matching the original intent of this entry.
    url: "/api/waterways?bbox=-74.5926,40.2508,-73.4074,41.1492",
    cacheTtl: 86400,
  },
};

export function OPTIONS() {
  return corsPreflightResponse();
}

/** Upstream GeoJSON geometry — only the coordinate pair is inspected. */
interface UpstreamGeometry {
  coordinates?: number[];
}

/**
 * Upstream GeoJSON feature. Property keys are open-ended and values are JSON
 * scalars — the property filter compares them with Number()/String(), which is
 * exactly how a scalar property is meant to be read.
 */
interface UpstreamFeature {
  type?: string;
  geometry?: UpstreamGeometry;
  properties?: Record<string, string | number | boolean | null | undefined>;
}

/**
 * NLNOG node entry as emitted by /api/nlnog. Coordinates may be absent or null;
 * such entries are dropped when the node list is converted to GeoJSON.
 */
interface NlnogNode {
  id?: number;
  hostname?: string;
  asn?: number;
  city?: string;
  country?: string;
  lat?: number | null;
  lon?: number | null;
}

/**
 * Upstream body: a GeoJSON FeatureCollection, or — for the nlnog_nodes
 * collection — an NLNOG `{nodes: [...]}` document. Malformed bodies surface
 * through the try/catch below, which keeps the 200 error-payload contract.
 */
interface UpstreamBody {
  features?: UpstreamFeature[];
  nodes?: NlnogNode[];
}

export async function GET(request: NextRequest, { params }: { params: Promise<{ id: string }> }) {
  const baseUrl = new URL(request.url).origin;
  const { id } = await params;
  const source = COLLECTION_SOURCES[id];

  if (!source) {
    return NextResponse.json(
      {
        code: "NotFound",
        description: `Collection '${id}' not found`,
      },
      { status: 404, headers: CORS_HEADERS },
    );
  }

  const url = new URL(request.url);
  // Non-numeric or non-positive limits fall back to the default — NaN would
  // otherwise flow into slice() and silently return an empty page.
  const rawLimit = parseInt(url.searchParams.get("limit") || "100", 10);
  const limit = Math.min(rawLimit >= 1 ? rawLimit : 100, 10000);
  const offset = parseInt(url.searchParams.get("offset") || "0", 10) || 0;
  const bbox = url.searchParams.get("bbox");
  const properties = url.searchParams.get("properties");

  try {
    const res = await fetch(source.url.startsWith("/") ? `${baseUrl}${source.url}` : source.url, {
      signal: AbortSignal.timeout(15000),
      headers: { "User-Agent": "OpenZenith/1.0" },
    });

    if (!res.ok) {
      return NextResponse.json(
        { error: `Upstream data source returned ${res.status}` },
        { status: 502, headers: CORS_HEADERS },
      );
    }

    const data = (await res.json()) as UpstreamBody;

    // Normalize to GeoJSON FeatureCollection
    // NLNOG returns { nodes: [...] } — convert to GeoJSON
    let features: UpstreamFeature[] = data.features || [];
    if (id === "nlnog_nodes" && data.nodes) {
      features = data.nodes
        .filter((n): n is NlnogNode & { lat: number; lon: number } => n.lat != null && n.lon != null)
        .map((n) => ({
          type: "Feature",
          geometry: { type: "Point", coordinates: [n.lon, n.lat] },
          properties: {
            id: n.id,
            hostname: n.hostname,
            asn: n.asn,
            city: n.city,
            country: n.country,
          },
        }));
    }

    // Bbox filtering: bbox=minLon,minLat,maxLon,maxLat
    if (bbox) {
      const [minLon, minLat, maxLon, maxLat] = bbox.split(",").map(Number);
      features = features.filter((f) => {
        const coords = f.geometry?.coordinates;
        if (!coords) return false;
        const lon = coords[0];
        const lat = coords[1];
        // A malformed bbox or a coordinate-pair-less geometry leaves a part
        // undefined; every comparison against undefined is false, so the
        // explicit guards below return exactly what the chained comparisons did.
        if (lon === undefined || lat === undefined) return false;
        if (minLon === undefined || minLat === undefined || maxLon === undefined || maxLat === undefined) {
          return false;
        }
        return lon >= minLon && lon <= maxLon && lat >= minLat && lat <= maxLat;
      });
    }

    // Simple property filtering: properties=field:value
    if (properties) {
      const filters = properties.split(",").map((f) => {
        // Default the value so a filter with no ":" separator degrades to a
        // non-matching string filter instead of crashing on undefined.
        // bounds: split always returns at least one element, so key is a string.
        const parts = f.split(":");
        const key = parts[0]!;
        const val = parts[1] ?? "";
        return { key, val };
      });

      features = features.filter((f) => {
        return filters.every(({ key, val }) => {
          const prop = f.properties?.[key];
          if (prop === undefined) return false;
          // Operator prefixes must be matched BEFORE parsing — parseFloat(">=5")
          // is NaN, which made every comparison branch unreachable.
          const ops: Array<[string, (a: number, b: number) => boolean]> = [
            [">=", (a, b) => a >= b],
            ["<=", (a, b) => a <= b],
            [">", (a, b) => a > b],
            ["<", (a, b) => a < b],
          ];
          for (const [prefix, cmp] of ops) {
            if (val.startsWith(prefix)) {
              const n = parseFloat(val.slice(prefix.length));
              return !isNaN(n) && cmp(Number(prop), n);
            }
          }
          // Plain value: numeric comparison when parseable
          const numVal = parseFloat(val);
          if (!isNaN(numVal)) return prop === numVal;
          // String comparison (case-insensitive)
          return String(prop).toLowerCase() === val.toLowerCase();
        });
      });
    }

    const total = features.length;
    const paged = features.slice(offset, offset + limit);
    const hasNext = offset + limit < total;

    // Build next/prev links
    const links = [
      {
        rel: "self",
        type: "application/geo+json",
        href: `${baseUrl}/api/collections/${id}/items?limit=${limit}&offset=${offset}`,
      },
      { rel: "collection", type: "application/json", href: `${baseUrl}/api/collections/${id}` },
      { rel: "root", type: "application/json", href: `${baseUrl}/api` },
    ];

    if (hasNext) {
      links.push({
        rel: "next",
        type: "application/geo+json",
        href: `${baseUrl}/api/collections/${id}/items?limit=${limit}&offset=${offset + limit}`,
      });
    }
    if (offset > 0) {
      links.push({
        rel: "prev",
        type: "application/geo+json",
        href: `${baseUrl}/api/collections/${id}/items?limit=${limit}&offset=${Math.max(0, offset - limit)}`,
      });
    }

    return NextResponse.json(
      {
        type: "FeatureCollection",
        features: paged,
        links,
        timeStamp: new Date().toISOString(),
        numberMatched: total,
        numberReturned: paged.length,
      },
      {
        headers: {
          "Content-Type": "application/geo+json",
          ...CORS_HEADERS,
          "Cache-Control": `public, max-age=${source.cacheTtl}`,
        },
      },
    );
  } catch (error) {
    const message = error instanceof Error ? error.message : "Failed to fetch features";
    return NextResponse.json({ error: message }, { status: 502, headers: CORS_HEADERS });
  }
}