# api/src/lib/openapi/base.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenZenith API",
    "version": "0.9.2",
    "description": "Free, fast, global elevation and geospatial API. Query any point on Earth for elevation data from NASA SRTM 30m, track flights, monitor weather, explore OpenStreetMap, and more. No API key required.",
    "contact": {
      "name": "OpenZenith",
      "url": "https://github.com/aliasfoxkde/OpenZenith"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current deployment (derived from request)"
    }
  ],
  "paths": {
...
    "/api/weather/warnings": {
      "get": {
        "summary": "Get weather warnings (NWS)",
        "description": "Returns NWS Watch/Warning/Advisory polygons from NOAA via ArcGIS. Optionally filter by bounding box geometry. Up to 500 features. 120-second cache.",
        "parameters": [
          {
            "name": "geometry",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Bounding box as xmin,ymin,xmax,ymax (WGS84) for spatial filtering",
            "example": "-125,25,-65,50"
          }
        ],
        "responses": {
          "200": {
            "description": "GeoJSON FeatureCollection of weather warnings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "GeoJSON FeatureCollection",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "FeatureCollection"
                    },
                    "features": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "description": "GeoJSON Feature with warning properties"
                      }
                    }
                  }
                }
              }
            }
          },
          "502": {
            "description": "Data fetch failed"
          }
        },
        "tags": ["Data"]
      }
    },
...
    "/api/nlnog": {
      "get": {
        "summary": "Get NLNOG Ring network nodes",
        "description": "Returns all NLNOG Ring measurement nodes worldwide with geographic coordinates, ASN, hostname, and city/country data. Approximately 743 nodes across 58 countries. Used for network infrastructure visualization. Cached for 1 hour.",
        "responses": {
          "200": {
            "description": "NLNOG node list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "nodes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 945
                          },
                          "hostname": {
                            "type": "string",
                            "example": "node.perth.ring.nlnog.net"
                          },
                          "asn": {
                            "type": "integer",
                            "example": 136557
                          },
                          "ipv4": {
                            "type": "string",
                            "example": "203.0.113.1"
                          },
                          "city": {
                            "type": "string",
                            "example": "Perth"
                          },
                          "country": {
                            "type": "string",
                            "example": "AU"
                          },
                          "lat": {
                            "type": "number",
                            "example": -31.86
                          },
                          "lon": {
                            "type": "number",
                            "example": 115.89
                          }
                        }
                      }
                    },
                    "count": {
                      "type": "integer",
                      "example": 743
                    }
                  }
                }
              }
            }
          },
          "502": {
            "description": "NLNOG API unavailable"
          }
        },
        "tags": ["Data"]
      }
    },
    "/api/bgp": {
      "get": {
        "summary": "BGP prefix lookup (NLNOG Looking Glass)",
        "description": "Queries the NLNOG Looking Glass API for BGP routing information about a given IP prefix. Returns AS paths, origin AS, and routing data. Cached for 5 minutes.",
        "parameters": [
          {
            "name": "prefix",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "IP prefix in CIDR notation",
            "example": "8.8.8.0/24"
          }
        ],
        "responses": {
          "200": {
            "description": "BGP routing data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "prefix": {
                      "type": "string",
                      "example": "8.8.8.0/24"
                    },
                    "data": {
                      "type": "object",
                      "description": "NLNOG Looking Glass response with AS path data"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing prefix parameter"
          },
          "502": {
            "description": "NLNOG Looking Glass unavailable"
          }
        },
        "tags": ["Network"]
      }
    },
...
    "/api/collections/{id}/items": {
      "get": {
        "summary": "Collection items (OGC Features)",
        "description": "Returns GeoJSON features for a specific collection with pagination and filtering.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Collection ID"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100
            },
            "description": "Items per page (max 10000)"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Start index"
          },
          {
            "name": "bbox",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Bounding box (minLon,minLat,maxLon,maxLat)"
          }
        ],
        "responses": {
          "200": {
            "description": "GeoJSON FeatureCollection"
          },
          "404": {
            "description": "Collection not found"
          }
        },
        "tags": ["Discovery"]
      }
    },
...
  },
...
}